requirements.txt
anyio==4.8.0
async-timeout==5.0.1
datasets==3.3.2
faiss-cpu==1.10.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.2
langchain-huggingface==0.1.2
networkx==3.4.2
numpy==2.2.3
openai==1.7.0
pandas==2.2.3
pillow==11.1.0
pyarrow==19.0.1
PyYAML==6.0.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.5.1
torch-max-mem==0.1.3
torch-ppr==0.0.8
tqdm==4.67.1
transformers==4.49.0